Add `:LspImplementation` mappings
`:LspDefinition` sometimes produces more than one result. When this is
the case the quickfix list is openend. This harms my workflow when
wanting to open the definition in a (v)split or tab. Although
`:LspImplementation`'s use-case is more broad than `:LspDefinition` (it
can go to but also find implementations) it can be a useful backup for
when I'm annoyed by `:LspDefinition`'s openeing of the quickfix list.